+++ patterns/lorem.php
@@ -1,9 +1,0 @@
-<?php
-/**
- * Title: Lorem
- * Slug: moare-steppe/lorem
- * Categories: moarepatterns
- * Description: A description of the pattern
- * Keywords: some words, separate by comas, etc
- */
-?>