+++ patterns/main-singular-without-margins.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Title: Main Singular Without Margins 
+ * Slug: moare-steppe/main-singular-without-margins
+ * Categories: moarepatterns
+ * Description: The main of singular without margins
+ * Keywords: Content
+ */
+?>
+
+<!-- wp:group {"tagName":"main","className":"ms-content"} -->
+<main class="wp-block-group ms-content" >
+	<!-- wp:group {"tagName":"article", "className":"ms-entry"} -->
+	<article class="wp-block-group ms-entry">
+		<!-- wp:post-content {"layout":{"type":"constrained"}} /-->
+	</article>
+	<!-- /wp:group -->
+</main>
+<!-- /wp:group -->