+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Moare Steppe Changelog
 
+## [1.0.2] - 2025-03-12
+### Added
+- Paragraph styles.
+- Patterns.
+- Stylesheets directory.
+
+### Changed
+- Page template hide default title.
+- Move main content to patterns.
+- Rename Header, Footer parts.
+- Some core block styles on theme.json
+
+### Remove
+- `page-without-title.html` file.
+
 ## [1.0.1] - 2025-01-21
 ### Changed
 - Update node packages.